Moved content of scripts-new from gulpfile.js to tools/build.js.
